+++ packages/polaris-viz/src/components/SimpleBarChart/tests/utilities.test.ts
@@ -43,106 +43,49 @@ const negativeSeries = {
 };
 
 describe('getLongestTrendIndicator()', () => {
-  it('returns a value for positive when the highest positive value has a trend indicator', () => {
-    const {positive, negative} = getLongestTrendIndicator(
-      [positiveSeries],
-      highestPositive,
-      0,
-    );
+  it('returns a value for positive when positive values have a trend indicator', () => {
+    const {positive, negative} = getLongestTrendIndicator([positiveSeries]);
 
-    expect(positive).toBeGreaterThan(0);
+    expect(positive).toStrictEqual(51);
     expect(negative).toBe(0);
   });
 
-  it('returns a value for negative when the lowest negative value has a trend indicator', () => {
-    const {positive, negative} = getLongestTrendIndicator(
-      [negativeSeries],
-      0,
-      lowestNegative,
-    );
+  it('returns a value for negative when negative values have a trend indicator', () => {
+    const {positive, negative} = getLongestTrendIndicator([negativeSeries]);
 
     expect(positive).toBe(0);
-    expect(negative).toBeGreaterThan(0);
+    expect(negative).toStrictEqual(51);
   });
 
-  it('returns values for both positive and negative when both the highest positive and lowest negative values have trend indicators', () => {
-    const {positive, negative} = getLongestTrendIndicator(
-      [positiveSeries, negativeSeries],
-      highestPositive,
-      lowestNegative,
-    );
+  it('returns values for both positive and negative when positive and negative values have trend indicators', () => {
+    const {positive, negative} = getLongestTrendIndicator([
+      positiveSeries,
+      negativeSeries,
+    ]);
 
-    expect(positive).toBeGreaterThan(0);
-    expect(negative).toBeGreaterThan(0);
-  });
-
-  it('returns 0 for positive when a lower value has a trend indicator but the highest positive value does not', () => {
-    const {positive, negative} = getLongestTrendIndicator(
-      [
-        {
-          ...positiveSeries,
-          metadata: {
-            trends: {
-              '0': trends['0'],
-            },
-          },
-        } as SimpleBarChartDataSeries,
-      ],
-      highestPositive,
-      0,
-    );
-
-    expect(positive).toBe(0);
-    expect(negative).toBe(0);
-  });
-
-  it('returns 0 for negative when a higher value has a trend indicator but the lowest negative value does not', () => {
-    const {positive, negative} = getLongestTrendIndicator(
-      [
-        {
-          ...negativeSeries,
-          metadata: {
-            trends: {
-              '0': trends['0'],
-            },
-          },
-        } as SimpleBarChartDataSeries,
-      ],
-      0,
-      lowestNegative,
-    );
-
-    expect(positive).toBe(0);
-    expect(negative).toBe(0);
+    expect(positive).toStrictEqual(51);
+    expect(negative).toStrictEqual(51);
   });
 
   it('returns 0 for both values if there are no trend indicators', () => {
-    const {positive, negative} = getLongestTrendIndicator(
-      [
-        {
-          ...positiveSeries,
-          metadata: {},
-        },
-      ],
-      highestPositive,
-      0,
-    );
+    const {positive, negative} = getLongestTrendIndicator([
+      {
+        ...positiveSeries,
+        metadata: {},
+      },
+    ]);
 
     expect(positive).toBe(0);
     expect(negative).toBe(0);
   });
 
   it('returns 0 for both values if metadata property is not present', () => {
-    const {positive, negative} = getLongestTrendIndicator(
-      [
-        {
-          ...positiveSeries,
-          metadata: undefined,
-        },
-      ],
-      highestPositive,
-      0,
-    );
+    const {positive, negative} = getLongestTrendIndicator([
+      {
+        ...positiveSeries,
+        metadata: undefined,
+      },
+    ]);
 
     expect(positive).toBe(0);
     expect(negative).toBe(0);